refactor(util): move bytes to string module
