+++ src/__tests__/modules/util.test.ts
@@ -11,7 +11,6 @@ import { concurrent, delay, disposeAll, wait } from '../../util'
 import { ansiparse, parseAnsiHighlights } from '../../util/ansiparse'
 import * as arrays from '../../util/array'
 import { filter } from '../../util/async'
-import bytes from '../../util/bytes'
 import * as color from '../../util/color'
 import { getSymbolKind } from '../../util/convert'
 import * as diff from '../../util/diff'
@@ -30,7 +29,7 @@ import * as positions from '../../util/position'
 import { executable, isRunning, runCommand, terminate } from '../../util/processes'
 import { convertProperties, Registry } from '../../util/registry'
 import { Sequence } from '../../util/sequence'
-import * as strings from '../../util/string'
+import bytes, * as strings from '../../util/string'
 import * as textedits from '../../util/textedit'
 import helper from '../helper'
 const createLogger = require('../../util/logger')
