+++ src/list/worker.ts
@@ -5,11 +5,10 @@ import { FuzzyMatch } from '../model/fuzzyMatch'
 import { IList, ListContext, ListItem, ListItemsEvent, ListItemWithScore, ListOptions, ListTask } from '../types'
 import { parseAnsiHighlights } from '../util/ansiparse'
 import { filter } from '../util/async'
-import bytes from '../util/bytes'
 import { patchLine } from '../util/diff'
 import { fuzzyMatch, getCharCodes } from '../util/fuzzy'
 import { Mutex } from '../util/mutex'
-import { smartcaseIndex } from '../util/string'
+import bytes, { smartcaseIndex } from '../util/string'
 import workspace from '../workspace'
 import Prompt from './prompt'
 const logger = require('../util/logger')('list-worker')
